修复实体世界传送
