+++ patches/net/minecraft/entity/Entity.java.patch
@@ -806,7 +806,7 @@
          }
      }
  
-@@ -2509,42 +2895,106 @@
+@@ -2509,42 +2895,87 @@
      @Nullable
      public Entity changeDimension(int dimensionIn)
      {
@@ -830,30 +830,11 @@
 +            // int i = this.dimension;
 +            // WorldServer worldserver = minecraftserver.getWorld(i);
 +            // WorldServer worldserver1 = minecraftserver.getWorld(dimensionIn);
-+            WorldServer exitWorld = null;
-+            if (this.dimension < CraftWorld.CUSTOM_DIMENSION_OFFSET) { // Plugins must specify exit from custom Bukkit worlds
-+                // Only target existing worlds (compensate for allow-nether/allow-end as false)
-+                for (WorldServer world : minecraftserver.worlds) {
-+                    if (world.dimension == dimensionIn) {
-+                        exitWorld = world;
-+                    }
-+                }
-+            }
++            WorldServer exitWorld = minecraftserver.getWorld(dimensionIn); // CatServer - allow hotload world
  
 -            if (i == 1 && dimensionIn == 1)
-+            BlockPos blockposition = null; // PAIL: CHECK
 +            Location enter = this.getBukkitEntity().getLocation();
-+            Location exit;
-+            if (exitWorld != null) {
-+                if (blockposition != null) {
-+                    exit = new Location(exitWorld.getWorld(), blockposition.getX(), blockposition.getY(), blockposition.getZ());
-+                } else {
-+                    exit = minecraftserver.getPlayerList().calculateTarget(enter, minecraftserver.getWorld(dimensionIn));
-+                }
-+            }
-+            else {
-+                exit = null;
-+            }
++            Location exit = exitWorld != null ? minecraftserver.getPlayerList().calculateTarget(enter, minecraftserver.getWorld(dimensionIn)) : null;
 +            boolean useTravelAgent = exitWorld != null && !(this.dimension == 1 && exitWorld.dimension == 1); // don't use agent for custom worlds or return from THE_END
 +
 +            TravelAgent agent = exit != null ? (TravelAgent) ((CraftWorld) exit.getWorld()).getHandle().getDefaultTeleporter() : org.bukkit.craftbukkit.CraftTravelAgent.DEFAULT; // return arbitrary TA to compensate for implementation dependent plugins
@@ -923,7 +904,7 @@
                  {
                      d0 = MathHelper.clamp(d0 * 8.0D, worldserver1.getWorldBorder().minX() + 16.0D, worldserver1.getWorldBorder().maxX() - 16.0D);
                      d1 = MathHelper.clamp(d1 * 8.0D, worldserver1.getWorldBorder().minZ() + 16.0D, worldserver1.getWorldBorder().maxZ() - 16.0D);
-@@ -2554,12 +3004,15 @@
+@@ -2554,12 +2985,15 @@
                  d1 = (double)MathHelper.clamp((int)d1, -29999872, 29999872);
                  float f = this.rotationYaw;
                  this.setLocationAndAngles(d0, this.posY, d1, 90.0F, 0.0F);
@@ -942,7 +923,7 @@
              this.world.profiler.endStartSection("reloading");
              Entity entity = EntityList.newEntity(this.getClass(), worldserver1);
  
-@@ -2567,7 +3020,8 @@
+@@ -2567,7 +3001,8 @@
              {
                  entity.copyDataFromOld(this);
  
@@ -952,7 +933,7 @@
                  {
                      BlockPos blockpos1 = worldserver1.getTopSolidOrLiquidBlock(worldserver1.getSpawnPoint());
                      entity.moveToBlockPosAndAngles(blockpos1, entity.rotationYaw, entity.rotationPitch);
-@@ -2576,19 +3030,28 @@
+@@ -2576,19 +3011,28 @@
                  {
                      entity.moveToBlockPosAndAngles(blockpos, entity.rotationYaw, entity.rotationPitch);
                  }
@@ -982,7 +963,7 @@
              return entity;
          }
          else
-@@ -2604,7 +3067,7 @@
+@@ -2604,7 +3048,7 @@
  
      public float getExplosionResistance(Explosion explosionIn, World worldIn, BlockPos pos, IBlockState blockStateIn)
      {
@@ -991,7 +972,7 @@
      }
  
      public boolean canExplosionDestroyBlock(Explosion explosionIn, World worldIn, BlockPos pos, IBlockState blockStateIn, float p_174816_5_)
-@@ -2717,6 +3180,11 @@
+@@ -2717,6 +3161,11 @@
  
      public void setCustomNameTag(String name)
      {
@@ -1003,7 +984,7 @@
          this.dataManager.set(CUSTOM_NAME, name);
      }
  
-@@ -2800,7 +3268,25 @@
+@@ -2800,7 +3249,25 @@
  
      public void setEntityBoundingBox(AxisAlignedBB bb)
      {
@@ -1030,7 +1011,7 @@
      }
  
      public float getEyeHeight()
-@@ -2901,6 +3387,218 @@
+@@ -2901,6 +3368,218 @@
          EnchantmentHelper.applyArthropodEnchantments(entityLivingBaseIn, entityIn);
      }
  
@@ -1249,7 +1230,7 @@
      public void addTrackingPlayer(EntityPlayerMP player)
      {
      }
-@@ -3069,7 +3767,7 @@
+@@ -3069,7 +3748,7 @@
          return SoundCategory.NEUTRAL;
      }
